osmo_verify_transcript: allow running/piping commands directly
To retrieve the VTY reference XML from 'show online-help', which we need for
the osmo-gsm-manuals, it is useful to directly print the output of a single VTY
command.

I have also always wanted to be able to just type CTRL commands and get their
results.

Add --command and --output arguments to run specific commands and pipe to
stdout instead of verifying a transcript file, in a way that works generically
for both VTY and CTRL interface.

The scripts' names are now a bit of a misnomer, to be tweaked separately.

Change-Id: I2c31a87fa255092bcd06da239651cc50174c8faa