osmo_interact_common: add '... !regex' wildcard
Sometimes it is desirable to match any lines, just make sure that a given regex
does *not* exist in the command reply.

For example, during handover config tests, to verify that no handover config
exists, one would have to match an entire 'show running-config' output, which
is subject to change whenever the VTY changes.

Add a wildcard '... !' which allows passing a regex that must not be
encountered, e.g.:

  OsmoBSC(config-net)# show running-config
  ...
  network
  ... !handover
   bts 0
  ...
    handover 1
  ...

This checks that there is no handover config on the network level while bts 0
does have handover config.

Change-Id: I7c1ebb2e7f059047903a53de26a0ec1ce7fa9b98