scripts/osmotestconfig: copy_config: no copytree
In Osmocom git repositories, we have doc/example directories (possibly
with subdirectories), which contain nothing but *.cfg files.

For these directories, the code I'm removing had no effect: it would
copy the whole directory tree of a directory containaing *.cfg files,
but for some reason ignore the *.cfg files. I guess we had a previous
directory structure before, where this made more sense. So essential it
is a mkdir, but this is also not needed since we do a mkdir just below
this.

Furthermore this code caused problems for the related osmo-sgsn patch,
where a *.cfg file is stored in the tests/ dir, because it has to be
slightly different than the file from doc/example. In this case, the
code would actually copy various files, such as Makefile.am, from the
tests directory. This causes problems in "make distcheck", because then
the file permissions are set so that these files cannot removed without
first changing the permissions again. So osmotestconfig.py fails when it
runs copy_config for the second time, trying to remove the temporary
directory.

Related: osmo-sgsn I309807ff0bc125d4653222b2b4ba69ded3bbff70
Change-Id: Ic312d546da1c21f68a80b6a188616ef9bc84f4c6